# docs/swagger.yaml
definitions:
...
  github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse:
    properties:
      code:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode'
      error:
        type: string
    type: object
...
  github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode:
    enum:
    - INVALID_CREDENTIALS
    - UNAUTHORIZED
    - FORBIDDEN
    - EMAIL_NOT_VERIFIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_ALREADY_EXISTS
    - INVALID_REQUEST
    - VALIDATION_FAILED
    - MISSING_REQUIRED_FIELD
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - OPERATION_FAILED
    - EMAIL_IS_ALREADY_IN_USE
    - EMAIL_IS_ALREADY_VERIFIED
    type: string
    x-enum-varnames:
    - ErrCodeInvalidCredentials
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeEmailNotVerified
    - ErrCodeResourceNotFound
    - ErrCodeResourceAlreadyExists
    - ErrCodeInvalidRequest
    - ErrCodeValidationFailed
    - ErrCodeMissingRequiredField
    - ErrCodeInternalServer
    - ErrCodeDatabaseError
    - ErrCodeOperationFailed
    - ErrCodeEmailIsAlreadyInUse
    - ErrCodeEmailIsAlreadyVerified
...
info:
  contact: {}
paths:
...
  /attachments/{id}/url:
    get:
      description: Generates a presigned URL to download an attachment. Access control
        is applied based on the attachment type.
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URL to download the attachment
          schema:
            type: string
        "400":
          description: Invalid attachment ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Authentication required or access denied
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "404":
          description: Attachment not found
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get attachment download URL
      tags:
      - attachments
...
swagger: "2.0"